lib/t5-bigbio-ner/pyproject.toml
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
include = '\.pyi?$'
line-length = 100
target-version = ['py36', 'py37', 'py38', 'py39']

[tool.coverage.report]
exclude_lines = [
  "if __name__ == .__main__.:",
  "pass",
  "pragma: no cover",
  "raise NotImplementedError",
  "raise TypeError",
  "raise ValueError"
]

[tool.isort]
known_local_folder = [
  're_verbalization',
  'tests'
]
line_length = 100
profile = 'black'

[tool.poetry]
authors = ["rian-t <rian.touchent@inria.fr>"]
description = "Toolkit for inference and finetuning of t5 models on multilingual NER using InstructionNER format"
name = "t5-bigbio-ner"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
datasets = "^2.11.0"
fastapi = "^0.95.1"
hydra-core = "^1.3.2"
python = "^3.9"
requests = "^2.28.2"
sentencepiece = "^0.1.98"
spacy = "^3.5.2"
torch = "^2.0.0"
tqdm = "^4.65.0"
transformers = "^4.28.1"
uvicorn = "^0.21.1"
wandb = "^0.15.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"

[tool.pytest.ini_options]
addopts = "--cov-report term-missing:skip-covered"
markers = [
  "serial",
  "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
